video.

+++ look-server/look-video/look-video-web/src/main/java/com/whoiszxl/cqrs/command/VideoSaveCommand.java
@@ -40,7 +40,7 @@ public class VideoSaveCommand implements Serializable {
     private String videoUrl;
 
     @ApiModelProperty("视频秒数")
-    private Float seconds;
+    private Integer seconds;
 
     @ApiModelProperty("视频宽度")
     private Integer width;
