+++ look-server/look-video/look-video-web/src/main/java/com/whoiszxl/cqrs/response/VideoResponse.java
@@ -49,7 +49,7 @@ public class VideoResponse implements Serializable {
     private String videoUrl;
 
     @ApiModelProperty("视频秒数")
-    private Float seconds;
+    private Integer seconds;
 
     @ApiModelProperty("视频宽度")
     private Integer width;
